Utility and minor bug fixes
