--- Project_FASTA/Project/FASTA/FASTA.depend
@@ -1,9 +1,0 @@
-# depslib dependency file v1.0
-1494219565 source:c:\myfiles\spbgu\proga\project_fasta\project\fasta\main.cpp
-	<iostream>
-	"fasta.cpp"
-
-1494219612 c:\myfiles\spbgu\proga\project_fasta\project\fasta\fasta.cpp
-	<cstdlib>
-	<cstdio>
-
