--- Project_FASTA/Project/FASTA/FASTA.layout
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<CodeBlocks_layout_file>
-	<FileVersion major="1" minor="0" />
-	<ActiveTarget name="Debug" />
-	<File name="fasta.cpp" open="1" top="1" tabpos="2" split="0" active="1" splitpos="0" zoom_1="3" zoom_2="0">
-		<Cursor>
-			<Cursor1 position="4446" topLine="2" />
-		</Cursor>
-		<Folding>
-			<Collapse line="4" />
-			<Collapse line="26" />
-			<Collapse line="41" />
-			<Collapse line="55" />
-			<Collapse line="85" />
-			<Collapse line="109" />
-			<Collapse line="140" />
-			<Collapse line="160" />
-			<Collapse line="189" />
-		</Folding>
-	</File>
-	<File name="main.cpp" open="1" top="0" tabpos="1" split="0" active="1" splitpos="0" zoom_1="3" zoom_2="0">
-		<Cursor>
-			<Cursor1 position="371" topLine="15" />
-		</Cursor>
-	</File>
-</CodeBlocks_layout_file>
