+++ Project_FASTA/Project/FASTA/main.cpp
@@ -18,22 +18,20 @@ int main()
     WrongFormat_wrongFirstSymbol
     UnknownSymbols
     */
-
     FASTA f1, f2;
     f1.setSequence("ACGTCAGACGACTACTACGATCGACGACTATCTATATATCTCGAGCTACGACTTACGAGCATCAT");
+    f1.print();
     cout << endl << endl;
     f2.setSequence("ATCGTATCGCGCTATAGGGCGGCAGACACACACGGTATGGACAGAGGCAGT");
+    f2.print();
     cout << endl << endl;
     f1.align(f2);
+    cout << endl << endl;
 
-    /*
-    FASTA f1;
     f1.read("../Samples/OctodonDegus_mRNA.fasta");
     f1.print();
     f1.save("../Samples/Saved/OctodonDegus_mRNA.fasta");
-    FASTA f2;
     f2.read("../Samples/Saved/OctodonDegus_mRNA.fasta");
     f2.print();
-    */
     return 0;
 }
